♻️ Refactoring code
+ Rename the `readParentIter()` function to `walkParentDir()`
+ Move functions and type declarations in the `src/utils/detect-package-manager.ts` file to the `src/utils/detect-package-manager` directory
